clk: qcom: Fix UBI32_PLL alpha value calculation
Huayra round_rate calculates the L and ALPHA values based on
rate requested and parent rate. When the remainder/quotient is
greater than 65535, the higher order bits are ignored during
left shift operation.

For UBI32_PLL, the requested rate was 1497.6MHz with parent rate
24MHz, whereas the obtained rate was 1488.032MHz since the quotient/
remainder variables could hold only 32bits during ALPHA calculation.

With this change, L and ALPHA values for ubi32_pll is set properly
to 0x3E and 0x6667 respectively for 1497.6MHz.

+++ drivers/clk/qcom/clk-alpha-pll.c
@@ -522,7 +522,7 @@ alpha_pll_huayra_round_rate(unsigned long rate, unsigned long prate,
 			    u32 *l, u32 *a)
 {
 	unsigned long remainder;
-	unsigned long quotient;
+	unsigned long long quotient;
 
 	quotient = rate;
 	remainder = do_div(quotient, prate);
@@ -533,7 +533,7 @@ alpha_pll_huayra_round_rate(unsigned long rate, unsigned long prate,
 		return rate;
 	}
 
-	quotient = remainder << PLL_HUAYRA_ALPHA_WIDTH;
+	quotient = (unsigned long long)remainder << PLL_HUAYRA_ALPHA_WIDTH;
 	remainder = do_div(quotient, prate);
 
 	if (remainder)
